Refactor observable references and fixes (see below.)
Changed Analysis to store the observable ids instead of observables. The
observable and observables properties still return Observable objects,
but they are loaded from the RootAnalysis instead. This removed the need
for the _load_observable_references function, and made it clear that
Observable and Analysis objects must exist inside of a RootAnalysis
object.

This also brought some bugs in the merging to the surface so there are
some changes to that as well.

Also changed JSON serialization so that by default the details
properties are stored in the JSON. Contrary to how the "1.0" version
worked, it is more common to include them in the JSON than not. The only
time they are not included is when Analysis objects are stored and
retrieved from the analysis tracking sub system. In those cases we do
*not* want to save/load the details, so an additional keyword arg of
exclude_analysis_details is added to the to_dict function that gets
propagated out to all to_dict() calls that chain out.