Refactored database code.
Moved the database code into ace/system/database since that is the only
place it is being used. Added testing to all of the support functions
defined in ace/system/database/__init__.py and moved the schema into the
same directory.

Adjusted ACESystem to include an initialize() function. Created
DatabaseACESystem as a base class for database-backed interfaces.

Using config interface for url and kwargs for SQLAlchemy parameters.