Improved content tracking.
Trying to figure out how file content gets uploaded along with the root
analysis requests or observable analysis results. Current strategy is
including the file content as part of the HTTP request for the
distributed model, and direct store for the non-distributed model.

Added a sync function executed on tracking root where we iterate all
file observables and make sure the meta tracks to the root. Should
refactor that later.