Modified usage of RootAnalysis.version. Added update_root_analysis API call.
The update_root_analysis() call will attempt to update the root if the
version matches, otherwise it returns an error code.

track_root_analysis() also modified to attempt to insert first, update
if it already exists, and return success or failure status.

The version property is updated by the interface API call if the insert
or update is successful.