Refactor to remove locking. More details below.
LOCKING

The distributed locking interface was removed from the design. There
were only two places where it was being used: root analysis and
requests.

The system was using a lock to cooridnate concurrent changes to root
objects. Now, the version field (now modified into a string) is used to
determine if an object changed. Attempts to update a root fail if the
version is found to be modified when the update is applied. Some simple
SQL syntax explains this:

UPDATE root SET blah = blah WHERE uuid = ? and version = ?

The version has to match. If it fails, then the caller is expected to
resolve the difference. This can be accomplished by using the new
update_and_save() function will acquires the most recent version of the
root, merges that into the root to be saved, and then again attempts to
save the root. This can happen any number of times, until the attempt to
save is successful.

For analysis requests, locking only applies to observable analysis
requests, for when an attempt is made to "link" one request to another.
Two additional interface functions to lock and unlock requests have been
added. It is up to the interface to implement the locks.

The analysis request lock is only acquired at the very start of the
processing of the request. Thus, any attempts to "modify" the request
must be made before the lock is acquired. The lock is never unlocked
because analysis requests are not re-used.

THREADED INTERFACE

A number of the interface implementations in the "threaded" system were
removed because they were redundant. The "database" system interface can
also be used in a threaded environment. The tests simply test that the
system as a whole works correctly. Having additional interfaces that do
the same thing (in a more difficult way) did not add any benefit.

EXPIRED ANALYSIS REQUEST PROCESSING

The implementation of this was moved into the interface, allowing for
more efficient code. An additional parameter was added to the call to
indicate what type should be checked for expired requests, rather than
always doing them all.