Refactored event data handling and ContentMeta.
ContentMetadata was moved into the data model. Added an additional class
to the data model called Event which contains the name of the event
along with any arguments. This allows the use of pydantic's JSON
serialization for that.

Refactored all the event handling to take an Event object as a single
parameter, or, an Event and Exception object for error handling.