Configuration refactor.
- added documentation for configuration settings
- refactored interface
- added data model for configuration settings
- added database-backed configuration settings
- added env-backed optional configuration settings
- bootstrap into database-backed configuration settings

The threaded version of the config interface was replaced by the
database version.

Notes

We use pydantic's JSON serialization to store configuration values. The
value column of the config table is actually the json of the
ace.data_model.ConfigurationSetting object (which includes the name and
the documentation) even though we actually store the name and
documentation in their own columns.

I did this so we can use pydantics JSON capabilities. Some improvement
could be made here.