Refactored interfaces into base classes.
I felt that having all of the interfaces in one class was a bit too much
for people new to the code base, and makes it difficult to navigate. All
of the interfaces now have base classes in ace/system/base, and the
ACESystem extends those.

Also refactored a few other things into their own modules, such as
logging.