Multiple bug fixes and testing design changes (see below.)
Separated the testing of the remote system from the local systems. The
remote system only actually implements a handful of the core api calls,
so it doesn't make sense to run the full suite of system tests against
it. All functions that are not actually a part of the remote api are
implemented to raise NotImplementedError now, so you can find cases
where you're accidentally calling it.

The module manager tests are what actually use the remote api.

Renamed submit_analysis_request to queue_analysis_request because
sometimes I mistakenly thought that was the function to use to send
requests to the core.

Fixed some random stuff I found by moving to running redis in a
container during the test.