Changes to support async operations inside of the CPU based analysis modules.
- Refactored the module functions into the class CPUTaskExecutor with
module-level wrapper functions.
- Changes to deal with issues when you have things happen on one thread
and event loop, and then things on another thread and event loop. The
redis connection pool object seemed to struggle with this.
- Changed the way AnalysisModuleManager objects are constructed. The
need a reference to a system they can use to interact with the core,
and also information to build the same system in the new processes that
get spawned when the manager's concurrency mode is process based.
- Refactored how the module testing works.