Changed extended_version from list to dict.
It makes a lot more sense to have this be a dictionary than a list
because you don't really know what the elements of the list are, and you
have to make sure you sort it to compare it.

With a dictionary you can basically assign names to values, and you also
don't have to worry about how it is compared.