Storage encryption support.
- everything in crypto is now async
- crypto functions were refactored into using async generators
- the file format of the encrypted files was changed to support iteration of
  data
- added EncryptionSettings to ACESystem
- added CONFIG_STORAGE_ENCRYPTION_ENABLED which controls storage encryption
- encryption is taken care of by the base storage class
- some refactoring had to happen so that sha256 and size could be computed as
  read
- attempts to read an unknown file raise an UnknownFileError now (instead of
  returning None)
- removed encryption "interface" from core system (not needed)

NOTES

There are a few things I'm focusing on here:
1) Iterating over data rather than processing it all at once.
2) Keeping the encryption abstract.
3) Making sure we do all our processing of file data in one pass.

The crypto functions were modified to work primarily with async generators.
This approach allows us to wrap async generators so that storage interface
implementations don't have to worry about encryption.

What this means is that the async generator that you receive will emit the
encrypted data of the source if encryption is enabled, so you just need to
store it. And likewise, when reading data, the bytes read from the async
generator will be the decrypted data.

The challenge with this approach is the computation of the sha256 and size of
the data. It has to be done as the original source is read, *before* we encrypt
it. So the i_store_content function now receives an additional parameter called
meta_computation of type MetaComputation. This object will contain the sha256
and size of the original content *after* the data has been read from the
source.

That means that after you've stored the encrypted data, you can use this
parameter to update the meta data with the correct sha256 and size of the
original unencrypted data.