Reintroduction of testing remote system.
Some of the things that can be done with the remote api are not covered by the
module testing. So it makes sense to reintroduce these tests. The catch is that
not all system api calls are implemented for the remote system (because there
is no use for them.)

So an additional marker called ace_remote can be added to tests in
tests/ace/system that will cause them to get executed when the system being
tested is the remote system. Otherwise they are skipped.

Also fixed some issues I found when I implemented parameter checking in
RemoteACESystem.